Updated `var` keywords to `let` and `const`, and a little reorganisation
Updating to `let` and `const` allows us to make it clearer on what variables will be reassigned further down the functions.

At the same time, tweaked the organisation of modules in `build.js` - moved the `argv` custom module down to where all the other custom modules are, and made sure they are correctly arranged in alphabetical order.