Restructured the boilerplate
In the previous structure, setting up a fresh boilerplate was a little cumbersome, as it involved having to delete the `.git` folder as well as having to update a large portion of the README so that it doesn't seem out of place when we push up the repository.

With this new structure, we no longer need to do the above two - we just need to make a fresh copy of the `dev` folder, then we are free to immediately initialise a new instance of Git in the folder; meanwhile, the README has been adjusted to be more generic - so all we have to do is update the `EDM Name` in it.