Major overhaul
1. Added in Nunjucks templating, to give us the option of handling
multiple EDMs within a single project.

2. CSS files are no longer removed immediately after they are processed
by InlineCSS - this is to give us the option of sharing a single CSS
file with multiple EDMs. (The generated CSS file will still be removed
from the build output, as we are not allowed to use external CSS files
within emails.)

3. Added a little more organization to the SCSS files, so that the
coding experience will be more similar to regular Metalsmith projects.

4. Added a setup script to simplify the project setup process involving
the update of the EDM name as well as the virtual folder name (similar
to the regular Metalsmith boilerplate.)

5. Removed stylelint from the build tool in favour of linting through
Sublime Text.

6. The `inlineCss` path in the YAML frontmatter will now be resolved
based on the file's current folder using Node's `path` module. This
means it's possible to use relative paths, which is useful if EDMs are
being organised into folders.